HCF-583 Address review comments
https://github.com/hpcloud/fissile/pull/159#pullrequestreview-4533018

Highlights:
- Move package-specific things into `builder/packages_image.go`
- Consistently call it "packages layer [[docker] image]"
